Manvita_GenerativeAI/requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
PyPDF2
python-docx
